requirements.txt
fastapi>=0.128.0
uvicorn[standard]>=0.40.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.10
PyJWT>=2.9.0
bcrypt>=4.2.0
pydantic>=2.12.0
pydantic-settings>=2.12.0
